clean the code and update it to be cell-based
The most notable change is the use of a contiguous storage for cells,
instead of the snake itself being the array. This saves quite a
substantial amount of space and time, I went with a 3-bit values
for each cell, making the whole SnakeContext struct 177 bytes, yet
it supports a 456 square area and multiple foods (in theory).